requirements.txt
numpy==1.26.4
pillow==10.3.0
opencv-python==4.10.0.84
matplotlib==3.9.2
tqdm==4.66.4
torch==2.3.1
torchvision==0.18.1
albumentations==1.4.10
scikit-image==0.24.0
scikit-learn==1.5.1
labelme==5.4.1
json5==0.9.14
ipython==8.25.0
ipywidgets==8.1.2